Cast alarm

+++ mpx_core_2020/mpx_core/modules/CommandHandler.c
@@ -554,7 +554,7 @@ int comhand(){
        memset(cmdBuffer, '\0', 100);
       sys_req(READ,DEFAULT_DEVICE,cmdBuffer,&bufferSize);
        println("");
-       addAlarm(message, cmdBuffer);
+       addAlarm(message, (int) cmdBuffer);
      }
    }
 //////////////////////////////////////////////////////////////////////////////////
